@@ -1,11 +1,17 @@
 import { StrictMode } from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import { App } from './App';
 import './index.css';
 
-ReactDOM.render(
+window.electron.onPong(() => {
+  console.log('pong!');
+});
+window.electron.ping();
+
+const container = document.getElementById('root') as HTMLElement;
+const root = createRoot(container);
+root.render(
   <StrictMode>
     <App />
-  </StrictMode>,
-  document.getElementById('root')
+  </StrictMode>
 );
