@@ -81,8 +81,8 @@ export function setMainMenu() {
   if (process.platform === 'darwin') {
     template.unshift(darwinMenuItem);
   } else {
-    if (!(template[0].submenu instanceof Menu)) {
-      template[0].submenu?.push(...[Separator, quitMenuItem]);
+    if (template[0].submenu instanceof Array) {
+      template[0].submenu.push(...[Separator, quitMenuItem]);
     }
   }
   Menu.setApplicationMenu(Menu.buildFromTemplate(template));
